Gestion Candidature avec les autorisation

+++ app/Http/Controllers/API/FormationController.php
@@ -4,11 +4,14 @@ namespace App\Http\Controllers\API;
 
 use App\Models\Formation;
 use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
 use GuzzleHttp\Promise\Create;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 
 class FormationController extends Controller
 {
+
+
     /**
      * Display a listing of the resource.
      */
@@ -34,6 +37,7 @@ class FormationController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('store', Formation::class);
         $validation = $request->validate([
             'libelle' => 'required|string',
             'duree' => 'required|string',
@@ -62,12 +66,35 @@ class FormationController extends Controller
         ]);
     }
 
+    public function formationCloturee()
+    {
+        $formations = Formation::where('cloturee', 1)->get();
+        return response()->json([
+            'message' => 'Liste des formation cloturées',
+            'Formations' => $formations
+        ]);
+    }
+
+
+    public function formationNonCloturee()
+    {
+        $formations = Formation::where('cloturee', 0)->get();
+        return response()->json([
+            'message' => 'Liste des formation non cloturées',
+            'Formations' => $formations
+        ]);
+    }
+
     /**
      * Show the form for editing the specified resource.
      */
     public function edit(Formation $formation)
     {
-        //
+        $formation->cloturee = 1;
+        $formation->save();
+        return response()->json([
+            'message' => 'La formation est cloturée'
+        ]);
     }
 
     /**
@@ -75,6 +102,7 @@ class FormationController extends Controller
      */
     public function update(Request $request, Formation $formation)
     {
+        Gate::authorize('update', Formation::class);
         $validation = $request->validate([
             'libelle' => 'required|string',
             'duree' => 'required|string',
@@ -101,6 +129,7 @@ class FormationController extends Controller
      */
     public function destroy(Formation $formation)
     {
+        Gate::authorize('delete', Formation::class);
         $formation->delete();
         return response()->json([
             'message' => 'Formation supprimer avec succes'
